net: bridge: move bridge files to folder
Moved bridge files to folder to manage. More bridge features
and files will be added in the future.
