+++ subsys/net/l2/ethernet/bridge/CMakeLists.txt
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+zephyr_library_include_directories(. ${ZEPHYR_BASE}/subsys/net/ip)
+
+zephyr_library_sources(bridge.c)
+zephyr_library_sources_ifdef(CONFIG_NET_ETHERNET_BRIDGE_SHELL bridge_shell.c)
